+++ components/tagger/apev2/apev2.cpp
@@ -733,7 +733,7 @@ Error BoCA::TaggerAPEv2::UpdateStreamInfo(const String &fileName, const Track &t
 
 			for (Int i = in.GetPos(); i < in.Size(); i += buffer.Size())
 			{
-				Int	 bytes = Math::Min(Int64(buffer.Size()), in.Size() - i);
+				Int	 bytes = Math::Min(buffer.Size(), in.Size() - i);
 
 				in.InputData(buffer, bytes);
 				out.OutputData(buffer, bytes);
@@ -779,7 +779,7 @@ Error BoCA::TaggerAPEv2::UpdateStreamInfo(const String &fileName, const Track &t
 
 			for (Int i = tagSize; i < in.Size(); i += buffer.Size())
 			{
-				Int	 bytes = Math::Min(Int64(buffer.Size()), in.Size() - i);
+				Int	 bytes = Math::Min(buffer.Size(), in.Size() - i);
 
 				in.InputData(buffer, bytes);
 				out.OutputData(buffer, bytes);
